Struct bholdus_currencies::module::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn transfer(
origin: OriginFor<T>,
dest: <T::Lookup as StaticLookup>::Source,
currency_id: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::CurrencyId,
amount: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance
) -> DispatchResult

Transfer some balance to another account under currency_id.

The dispatch origin for this call must be Signed by the transactor.

pub fn transfer_native_currency(
origin: OriginFor<T>,
dest: <T::Lookup as StaticLookup>::Source,
amount: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

Transfer some native currency to another account.

The dispatch origin for this call must be Signed by the transactor.